refactor: rename _common.py to session_handlers.py for better clarity
- Renamed _common.py to session_handlers.py
- Updated imports in _audio_handler.py
- Updated imports in _text_handler.py
- Updated imports in message_handlers.py
- Improved code discoverability for team members
