+++ app/roleplaying/handlers/_audio_handler.py
@@ -24,7 +24,7 @@ from app.roleplaying.core.session_state_manager import (
     SessionMessageHandler,
     SessionAudioHandler,
 )
-from app.roleplaying.handlers._common import (
+from app.roleplaying.handlers.session_handlers import (
     _send_error,
     _check_turn_limit,
     _evaluate_feedback,
